+++ test_predsim.py
@@ -28,22 +28,39 @@ TESTFILES_DIR = os.path.join(
     os.path.dirname(os.path.realpath(__file__)), 'test_files')
 
 
+with open(os.path.join(TESTFILES_DIR, 'expected-hky-3.nex'), 'r') as fo:
+    EXP_NEX_3 = fo.read()
+
+with open(os.path.join(TESTFILES_DIR, 'expected-hky-1.nex'), 'r') as fo:
+    EXP_NEX_1 = fo.read()
+
+with open(os.path.join(TESTFILES_DIR, 'expected-hky-3.phy'), 'r') as fo:
+    EXP_PHY_3 = fo.read()
+
+with open(os.path.join(TESTFILES_DIR, 'expected-hky-1.phy'), 'r') as fo:
+    EXP_PHY_1 = fo.read()
+
+
+def versiontuple(v):
+    return tuple(map(int, (v.split("."))))
+
+
 def seqgen_status(path):
     """
     Return True if Seq-Gen executable is working,
     otherwise return False.
     """
-    f = open(os.devnull, 'w')
+    fo = open(os.devnull, 'w')
     try:
         subprocess.check_call(
-            SEQGEN_PATH, stdout=f, stderr=subprocess.STDOUT)
+            SEQGEN_PATH, stdout=fo, stderr=subprocess.STDOUT)
         status = True
     except subprocess.CalledProcessError:
         status = False
     except OSError:
         status = False
     finally:
-        f.close()
+        fo.close()
     return status
 
 
@@ -53,53 +70,28 @@ seqgen_required = pytest.mark.skipif(
 
 class TestReadTreeFile():
 
-    t_file_string = (
-        """
-        #NEXUS
-        [ID: 9409050143]
-        [Param: tree]
-        begin trees;
-            translate
-            1 t1,
-            2 t2,
-            3 t3,
-            4 t4;
-        tree rep.1 = ((1:0.1,2:0.1):0.1,3:0.1,4:0.1);
-        tree rep.2 = ((1:0.1,2:0.1):0.1,3:0.1,4:0.1);
-        end;""")
-
-    def test_read_tree_file(self, tmpdir):
-        f = tmpdir.join('t-file.txt')
-        f.write(self.t_file_string)
-        tree_list = read_tfile(str(f.dirpath('t-file.txt')))
-        assert len(tree_list) == 2
+    def test_read_tree_file(self):
+        tree_list = read_tfile(os.path.join(TESTFILES_DIR, 'data-hky.t'))
+        assert len(tree_list) == 3
 
     def test_read_empty_tree_file(self, tmpdir):
-        f = tmpdir.join('empty-t-file.txt')
-        f.write('')
+        fo = tmpdir.join('empty-t-file.txt')
+        fo.write('')
         with pytest.raises(ValueError):
-            read_pfile(str(f.dirpath('empty-t-file.txt')))
+            read_pfile(str(fo.dirpath('empty-t-file.txt')))
 
 
 class TestReadParameterFile():
 
-    p_file_string = (
-        '[ID: 1234567890]\n'
-        'Gen\tLnL\tLnPr\tTL\tkappa\tpi(A)\tpi(C)\tpi(G)\tpi(T)\talpha\n'
-        '0\t-5\t50\t0.5\t1\t0.25\t0.25\t0.25\t0.25\t1\n'
-        '500\t-5\t50\t0.5\t1\t0.25\t0.25\t0.25\t0.25\t1')
-
-    def test_read_parameter_file(self, tmpdir):
-        f = tmpdir.join('p-file.txt')
-        f.write(self.p_file_string)
-        p_dicts = read_pfile(str(f.dirpath('p-file.txt')))
-        assert len(p_dicts) == 2
+    def test_read_parameter_file(self):
+        p_dicts = read_pfile(os.path.join(TESTFILES_DIR, 'data-hky.p'))
+        assert len(p_dicts) == 3
 
     def test_read_empty_parameter_file(self, tmpdir):
-        f = tmpdir.join('empty-p-file.txt')
-        f.write('')
+        fo = tmpdir.join('empty-p-file.txt')
+        fo.write('')
         with pytest.raises(ValueError):
-            read_pfile(str(f.dirpath('empty-p-file.txt')))
+            read_pfile(str(fo.dirpath('empty-p-file.txt')))
 
 
 class TestKappaConversion():
@@ -144,18 +136,18 @@ class TestSingleSimulation():
         result = simulate_matrix(self.tree, seqgen_path=SEQGEN_PATH)
         assert len(result.char_matrix) == 4
         assert result.char_matrix.sequence_size == 1000
-        assert result.tree == self.tree
-        assert 'HKY' in result.command_line
+        assert result.tree == self.tree.as_string('newick') + '\n'
+        assert 'HKY' in result.command
 
     def test_gtr(self):
         result = simulate_matrix(
             self.tree, general_rates=self.rates, seqgen_path=SEQGEN_PATH)
-        assert 'GTR' in result.command_line
+        assert 'GTR' in result.command
 
     def test_ti_tv(self):
         result = simulate_matrix(self.tree, ti_tv=1, seqgen_path=SEQGEN_PATH)
-        assert 'HKY' in result.command_line
-        assert ' -t1 ' in result.command_line
+        assert 'HKY' in result.command
+        assert ' -t1 ' in result.command
 
     def test_ti_tv_and_gtr(self):
         with pytest.raises(ValueError):
@@ -166,13 +158,13 @@ class TestSingleSimulation():
     def test_gamma(self):
         result = simulate_matrix(
             self.tree, gamma_shape=2, seqgen_path=SEQGEN_PATH)
-        assert ' -a2 ' in result.command_line
+        assert ' -a2 ' in result.command
 
     def test_gamma_shape_and_cats(self):
         result = simulate_matrix(
             self.tree, gamma_shape=2, gamma_cats=5, seqgen_path=SEQGEN_PATH)
-        assert ' -a2 ' in result.command_line
-        assert ' -g5 ' in result.command_line
+        assert ' -a2 ' in result.command
+        assert ' -g5 ' in result.command
 
     def test_gamma_cats(self):
         with pytest.raises(ValueError):
@@ -257,9 +249,8 @@ class TestArgumentParser():
             with tempfile.NamedTemporaryFile() as t_file:
                 with tempfile.NamedTemporaryFile() as commands_file:
                     parse_args([
-                        '-l100', '-s1', '-g4',
-                        '--commands-file', commands_file.name,
-                        p_file.name, t_file.name])
+                        '-l100', '-s1', '-g4', '--commands-file',
+                        commands_file.name, p_file.name, t_file.name])
 
     def test_is_file(self):
         with tempfile.NamedTemporaryFile() as tmp:
@@ -275,6 +266,8 @@ class TestMain():
 
     outfile = tempfile.NamedTemporaryFile('w')
 
+    phylip_out = '4 2\nt1  GT\nt2  GT\nt3  CT\nt4  GT\n'
+
     def test_args_help(self):
         with pytest.raises(SystemExit):
             main(['-h'])
@@ -285,25 +278,55 @@ class TestMain():
 
     def test_hky(self):
         main([
-            '-l', '10',
-            os.path.join(TESTFILES_DIR, 'anolis_hky.p'),
-            os.path.join(TESTFILES_DIR, 'anolis_hky.t')])
+            '-l', '2',
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+
+    def test_hky_seeds_file(self, capsys):
+        main([
+            '-l', '2',
+            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds-3.txt'),
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+        captured = capsys.readouterr()
+        if versiontuple(pytest.__version__) >= (3, 3, 0):
+            assert captured.out == EXP_NEX_3
+            assert captured.err == ''
+
+    def test_hky_skip(self, capsys):
+        main([
+            '-l', '2', '-s', '2',
+            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds-1.txt'),
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+        captured = capsys.readouterr()
+        if versiontuple(pytest.__version__) >= (3, 3, 0):
+            assert captured.out == EXP_NEX_1
+            assert captured.err == ''
 
     def test_hky_commands_file(self):
         main([
-            '-l', '10', '--commands-file', self.outfile.name,
-            os.path.join(TESTFILES_DIR, 'anolis_hky.p'),
-            os.path.join(TESTFILES_DIR, 'anolis_hky.t')])
+            '-l', '2', '-s', '2', '--commands-file', self.outfile.name,
+            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds-1.txt'),
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+        assert os.path.isfile(self.outfile.name)
 
-    def test_hky_seeds_file(self):
+    def test_hky_trees_file(self):
         main([
-            '-l', '10',
-            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds.txt'),
-            os.path.join(TESTFILES_DIR, 'anolis_hky.p'),
-            os.path.join(TESTFILES_DIR, 'anolis_hky.t')])
+            '-l', '2', '-s', '2', '--trees-file', self.outfile.name,
+            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds-1.txt'),
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+        assert os.path.isfile(self.outfile.name)
 
-    def test_hky_phylip(self):
+    def test_hky_phylip(self, capsys):
         main([
-            '-l', '10', '-o', 'phylip',
-            os.path.join(TESTFILES_DIR, 'anolis_hky.p'),
-            os.path.join(TESTFILES_DIR, 'anolis_hky.t')])
+            '-l', '2', '-s', '2', '-o', 'phylip',
+            '--seeds-file', os.path.join(TESTFILES_DIR, 'seeds-1.txt'),
+            os.path.join(TESTFILES_DIR, 'data-hky.p'),
+            os.path.join(TESTFILES_DIR, 'data-hky.t')])
+        captured = capsys.readouterr()
+        if versiontuple(pytest.__version__) >= (3, 3, 0):
+            assert captured.out == EXP_PHY_1
+            assert captured.err == ''